seems ok, but slow...need to fix the bug of react server-client rendering and then test it...
